Simplify Halstead metric
